+++ lib/src/components/widgets/pdf_watermark.dart
@@ -1,3 +1,5 @@
+// ignore_for_file: deprecated_member_use
+
 import 'dart:ui';
 
 import 'package:syncfusion_flutter_pdf/pdf.dart';